+++ src/main/java/roomescape/config/ReservationConfig.java
@@ -3,22 +3,23 @@ package roomescape.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.repository.ReservationRepository;
-import roomescape.repository.ReservationRepositoryV1;
+import roomescape.repository.ReservationRepositoryImpl;
 import roomescape.service.ReservationService;
-import roomescape.service.ReservationServiceV1;
+import roomescape.service.ReservationServiceImpl;
 
 @Configuration
-@Profile("v1")
-public class ReservationConfigV1 {
+@Profile("v2")
+public class ReservationConfig {
 
     @Bean
-    public ReservationRepository reservationRepository() {
-        return new ReservationRepositoryV1();
+    public ReservationRepository reservationRepository(JdbcTemplate jdbcTemplate) {
+        return new ReservationRepositoryImpl(jdbcTemplate);
     }
 
     @Bean
     public ReservationService reservationService(ReservationRepository reservationRepository) {
-        return new ReservationServiceV1(reservationRepository);
+        return new ReservationServiceImpl(reservationRepository);
     }
 }
